V1

+++ javaTraining/src/dao/CompanyDAO.java
@@ -7,5 +7,5 @@ import model.Company;
 
 public interface CompanyDAO {
 	public ArrayList<Company> findAllCompanies () throws DAOException;
-	public Company findByIdCompany (int id) throws DAOException;
+	public Company findByIdCompany(long id) throws DAOException;
 }
